requirements.txt
google-generativeai
google-auth
google-api-python-client
supabase
python-dotenv
importlib-metadata
beautifulsoup4
requests
